core(ci): Docker tag not dependent of it on PR (#2894)
- On PR like on master, if we haven't any it job docker-tag job
should be triggered.
- Trick used : By default a if statement which not contain any status function
will trigger a success() if our condition is true.
Problem : We mix a needs statement and in that case all needed task must be successful
to trigger the job with a success...
Solution : Mix in if statement the always() status function.
Seems a bit hacky but works as expected !
GitOrigin-RevId: 49810f849533a6912040692661eb4b2420aa2cb9