fix(ci): use same paths for pull/push npm cache
Paths are used to determine which cache to pull.
Using different paths for push and pull, will make jobs expecting to find cache (eg: precheck) fail retrieving it for the key and would push it at the end of the job, resulting in an empty cache (because no deps are installed in such steps).

GitOrigin-RevId: 768ba568fee2daa2040baf3535e971a93c79c06d